# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project PM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks/{taskId}/notes:
    get:
      operationId: listTaskNotes
      tags: [tasks]
      summary: List notes (including status-change notes) for a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskNote"
    post:
      operationId: createTaskNote
      tags: [tasks]
      summary: Add a note to a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskNote"
...
  /share-links:
    get:
      operationId: listShareLinks
      tags: [shares]
      summary: List share links for a given resource
      parameters:
        - in: query
          name: resourceType
          required: true
          schema: { type: string, enum: [project, task, document] }
        - in: query
          name: resourceId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Active and inactive links for the resource
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ShareLink" }
    post:
      operationId: createShareLink
      tags: [shares]
      summary: Create a new share link
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShareLinkInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShareLink" }
  /share-links/{id}:
    delete:
      operationId: revokeShareLink
      tags: [shares]
      summary: Revoke a share link
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Revoked }
  /share-links/{id}/email:
    post:
      operationId: emailShareLink
      tags: [shares]
      summary: Email an existing share link to one or more recipients
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShareLinkEmailInput" }
      responses:
        "200":
          description: Email accepted by the provider
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShareLinkEmailResult" }
        "400": { description: No valid recipients or link is revoked/expired }
        "404": { description: Share link not found }
        "503": { description: Email is not configured on this server }
...
components:
  schemas:
...
    ShareLinkEmailInput:
      type: object
      required: [recipients]
      properties:
        recipients:
          type: array
          minItems: 1
          items: { type: string }
          description: Email addresses to send the share link to.
        message:
          type: string
          maxLength: 2000
          description: Optional personal note included in the email body.
    ShareLinkEmailResult:
      type: object
      required: [sent, recipients]
      properties:
        sent: { type: integer }
        recipients:
          type: array
          items: { type: string }
    ShareLinkInput:
      type: object
      required: [resourceType, resourceId]
      properties:
        resourceType:
          type: string
          enum: [project, task, document]
        resourceId: { type: integer }
        expiresAt:
          type: string
          description: ISO timestamp. Omit for no expiry.
    ShareLink:
      type: object
      required: [id, token, resourceType, resourceId, url, createdAt, active]
      properties:
        id:           { type: integer }
        token:        { type: string }
        resourceType: { type: string, enum: [project, task, document] }
        resourceId:   { type: integer }
        url:          { type: string, description: "Absolute URL to the public share page." }
        createdBy:    { type: integer }
        createdAt:    { type: string }
        expiresAt:    { type: string }
        revokedAt:    { type: string }
        active:       { type: boolean }
...
    TaskNote:
      type: object
      required: [id, taskId, body, createdAt]
      properties:
        id:           { type: integer }
        taskId:       { type: integer }
        authorId:     { type: ["integer", "null"] }
        authorName:   { type: ["string", "null"] }
        body:         { type: string }
        statusBefore: { type: ["string", "null"] }
        statusAfter:  { type: ["string", "null"] }
        createdAt:    { type: string }
    TaskNoteInput:
      type: object
      required: [body]
      properties:
        body:         { type: string, minLength: 1 }
        authorName:   { type: string }
        authorId:     { type: integer }

    # ── Comment ─────────────────────────────────────────────────────────────
...